New test radio with parametr
