--- src/test/java/RadioTestWithParametr.java
@@ -1,159 +1,0 @@
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class RadioTestWithParametr {
-
-    @Test
-    public void testFullSizeStations() {   // Граници номеров полного диапазона радиостанций (метод size)
-        RadioWithParametr radio = new RadioWithParametr(8);
-        Assertions.assertEquals(0, radio.getMinCarrentStation());
-        Assertions.assertEquals(0, radio.getCarrentStation());
-        Assertions.assertEquals(8, radio.getMaxCarrentStation());
-    }
-
-    @Test
-    public void testMinMaxStations() { // Установка рабочего диапазона станций внутри допустимых значений их номеров
-        RadioWithParametr radio = new RadioWithParametr(2, 7);
-        Assertions.assertEquals(2, radio.getMinCarrentStation());
-        Assertions.assertEquals(2, radio.getCarrentStation());
-        Assertions.assertEquals(7, radio.getMaxCarrentStation());
-    }
-
-    @Test
-    public void test0Stations() {      // Минимальное число и номер радиостанции в диапазоне
-        RadioWithParametr radio = new RadioWithParametr(0);
-        Assertions.assertEquals(0, radio.getMinCarrentStation());
-        Assertions.assertEquals(0, radio.getCarrentStation());
-        Assertions.assertEquals(0, radio.getMaxCarrentStation());
-    }
-
-    @Test
-    public void testPrevStationsLouHi() {  // Перехода через нижнее значение номера радиостанции ввниз
-        RadioWithParametr radio = new RadioWithParametr(0, 9);
-        radio.prev();
-        int actual = radio.getCarrentStation();
-        int expected = 9;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testPrevStationsHiLou() {  // Переход через верхнее значение номера радиостанции вверх
-        RadioWithParametr radio = new RadioWithParametr(9);
-        radio.setCarrentStation(9);
-        radio.next();
-        int actual = radio.getCarrentStation();
-        int expected = 0;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testNextStationsInside() { // Переключение с указанного номера вверх на границу указанного диапазона
-        RadioWithParametr radio = new RadioWithParametr(3);
-        radio.setCarrentStation(2);
-        radio.next();
-        int actual = radio.getCarrentStation();
-        int expected = 3;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testPrevStationsOuteside() { // Переключение с указанного номера вниз на границу указанного диапазона
-        RadioWithParametr radio = new RadioWithParametr(3);
-        radio.setCarrentStation(5);
-        radio.prev();
-        int actual = radio.getCarrentStation();
-        int expected = 3;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testNextStationsOuteside() { // Переключение с указанного номера за границей указанного диапазона вниз
-        RadioWithParametr radio = new RadioWithParametr(3);
-        radio.setCarrentStation(5);
-        radio.next();
-        int actual = radio.getCarrentStation();
-        int expected = 1;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testNextStationsOutesideMax() { // Переключение с указанного номера за границей указанного диапазона вниз
-        RadioWithParametr radio = new RadioWithParametr(9);
-        radio.setCarrentStation(0);
-        radio.prev();
-        int actual = radio.getCarrentStation();
-        int expected = 9;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testPrevStationsOutesideMin() { // Переключение с "невозможного" номера за границей
-        RadioWithParametr radio = new RadioWithParametr(-1, 5);            // указанного диапазона вниз (проверка покрытия тестов)
-        radio.setCarrentStation(0);
-        radio.prev();
-        int actual = radio.getCarrentStation();
-        int expected = -1;
-        Assertions.assertEquals(expected, actual);
-    }
-
-
-    @Test
-    public void testUpVolume_0up() {
-        RadioWithParametr radio = new RadioWithParametr(20);
-        radio.setCarrentVolume(0);
-        radio.up();
-        int actual = radio.getCarrentVolume();
-        int expected = 1;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testUpVolume_9Up() {
-        RadioWithParametr radio = new RadioWithParametr(20);
-        radio.setCarrentVolume(101);
-        radio.up();
-        int actual = radio.getCarrentVolume();
-        int expected = 1;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testUpVolume_100up() {
-        RadioWithParametr radio = new RadioWithParametr(20);
-        radio.setCarrentVolume(100);
-        radio.up();
-        int actual = radio.getCarrentVolume();
-        int expected = 100;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testDowneVolume_0down() {
-        RadioWithParametr radio = new RadioWithParametr(30);
-        radio.setCarrentVolume(0);
-        radio.down();
-        int actual = radio.getCarrentVolume();
-        int expected = 0;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testDowneVolume_100down() {
-        RadioWithParametr radio = new RadioWithParametr(30);
-        radio.setCarrentVolume(100);
-        radio.down();
-        int actual = radio.getCarrentVolume();
-        int expected = 99;
-        Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void testDowneVolume_negdown() {
-        RadioWithParametr radio = new RadioWithParametr(30);
-        radio.setCarrentVolume(-1);
-        radio.down();
-        int actual = radio.getCarrentVolume();
-        int expected = 0;
-        Assertions.assertEquals(expected, actual);
-    }
-}
